refactor: qualify walk_nodes docstring for DAGs, deduplicate test helper
Qualify the walk_nodes docstring to note that parent-before-descendant
order only holds for tree-shaped expressions; shared (DAG) nodes may
appear before a non-matching ancestor.

Move the duplicated _replay_rebuild helper from both test files into
conftest.py.
